backend/api/requirements.txt
# ROSETTA API Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM inference
llama-cpp-python==0.2.20

# Configuration
pyyaml==6.0.1

# File validation (optional)
python-magic==0.4.27

# OCR dependencies (from main project)
paddlepaddle==2.5.1
paddleocr==2.7.0
pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
